Changed paths: M      cg_vbm8_update.m M      cg_spmF2x.m M      cg_spmT2x.m
(1) Added option in update function to skip update.
(2) Changed datatype in cg_spmT2x and cg_spmF2x to float32.


git-svn-id: svn://141.35.69.214/cat12@309 9f44b166-36c4-483c-8295-4af200601734

+++ cg_spmT2x.m
@@ -366,9 +366,9 @@ for i=1:size(P,1)
        VO = Vspm;
        VO.fname = out;
        if spm2
-            VO.dim(4) = spm_type('int16');
+            VO.dim(4) = spm_type('float32');
        else
-            VO.dt = [spm_type('int16') spm_platform('bigend')];
+            VO.dt = [spm_type('float32') spm_platform('bigend')];
        end
        spm_write_vol(VO,Y);
